# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Outreach Server
  version: 0.1.0
  description: |
    The `/v1` surface of the outreach engine — the real API.

    **This document covers `/v1` only.** The 110 legacy endpoints served
    alongside it are a compatibility shim (`src/api/compat/`), deleted in P12,
    and are inventoried in the extraction plan's Appendix A rather than here.
    Documenting them would give them a life they are not meant to have. Every
    compat response carries `Deprecation: true` and a `Link` header naming its
    successor below.

    ## Tenancy

    Every request is tenant-scoped, and **the tenant is never read from the body
    or the query string** — it comes from the gateway-forwarded identity headers.
    A body field named `tenantId` or `medspaId` is ignored, and the MCP router
    logs a warning when it sees one.

    ## Authentication

    The gateway terminates user auth and forwards identity headers. Three
    surfaces are deliberately reachable without them, each for a stated reason:

    | Path | Why |
    |---|---|
    | `GET /metrics` | Prometheus scrapes with no headers |
    | `GET /mcp/tools`, `GET /mcp/health` | tera-orchestrator builds its tool registry at its own startup. Discovery is schema-only; every tool that touches data is authenticated. |
    | `POST /v1/webhooks/*` | a provider callback has no headers. The signature is the credential. |
...

  license:
    name: Proprietary
servers:
  - url: https://{host}/api/communication
...
      host:
...
security:
  - gatewayIdentity: []
paths:
...
  /v1/content/generate:
    post:
      operationId: postContentGenerate
      tags: [content]
      summary: Generate a draft from a prompt pack
      description: |
        `aiConfidence` is a **deterministic composite** of things the engine can
        observe — context completeness, less a penalty per lint warning — not
        the model's self-assessment. It is a heuristic, not a probability, and
        the approval `threshold` mode that consumes it ships disabled.
      security: [{ gatewayIdentity: [outreach:send] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [promptPackKey, channel]
              properties:
                promptPackKey: { type: string }
                channel: { $ref: '#/components/schemas/Channel' }
                playbookKey: { type: string }
                playbookGoal: { type: string }
                data: { $ref: '#/components/schemas/RenderContext' }
                overrides:
                  type: object
                  properties:
                    tone: { type: string }
                    language: { type: string }
                    model: { type: string }
      responses:
        '200':
          description: The draft.
          content: { application/json: { schema: { $ref: '#/components/schemas/GeneratedDraft' } } }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /v1/assets:
    get:
      operationId: getAssets
      tags: [assets]
      summary: List this tenant's assets
      parameters:
        - { name: kind, in: query, schema: { type: string, enum: [image, document, generated] } }
        - { $ref: '#/components/parameters/Limit' }
        - { name: offset, in: query, schema: { type: integer } }
      responses:
        '200':
          description: The assets.
          content:
            application/json:
              schema:
                type: object
                properties:
                  assets: { type: array, items: { $ref: '#/components/schemas/Asset' } }
                  count: { type: integer }
    post:
      operationId: postAssets
      tags: [assets]
      summary: Upload an asset
      description: |
        `multipart/form-data` with a single `file` part, bounded by
        `ASSET_MAX_BYTES` (10 MB by default).

        The stored key is **assigned by the service** — `<tenantId>/<kind>/<uuid>.<ext>`.
        The filename you send is recorded in `metadata.filename` for display and
        never reaches a path.
      security: [{ gatewayIdentity: [outreach:templates:write] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        '201':
          description: The stored asset.
          content: { application/json: { schema: { $ref: '#/components/schemas/Asset' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '501':
          description: No storage adapter is configured.
  /v1/assets/generate:
    post:
      operationId: postAssetsGenerate
      tags: [assets]
      summary: Generate an image asset
      description: |
        **Answers 501 in every shipped configuration.** The `ImageProvider` port
        is declared (`src/ports/image.ts`) and the engine registers no adapter,
        because the service this was extracted from could not generate an image
        either — `AIService.generateImage` threw unconditionally. See D92.

        A deployment with an image model registers an adapter in the composition
        root and this begins working; no other code changes.
      security: [{ gatewayIdentity: [outreach:templates:write] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [prompt]
              properties:
                prompt: { type: string }
                size:
                  type: string
                  enum: ['256x256', '512x512', '1024x1024', '1792x1024', '1024x1792']
                style: { type: string, enum: [natural, vivid] }
                count: { type: integer, minimum: 1, maximum: 4 }
                model: { type: string }
      responses:
        '201':
          description: The generated assets.
          content:
            application/json:
              schema:
                type: object
                properties:
                  assets: { type: array, items: { $ref: '#/components/schemas/Asset' } }
                  count: { type: integer }
        '501':
          description: No image model is configured.
  /v1/assets/{id}:
    get:
      operationId: getAssetsById
      tags: [assets]
      summary: Get one asset
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: The asset.
          content: { application/json: { schema: { $ref: '#/components/schemas/Asset' } } }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      operationId: deleteAssetsById
      tags: [assets]
      summary: Delete an asset
      description: |
        Removes the row, then the stored object. That order is deliberate: an
        orphaned object costs storage, an orphaned row costs a broken image in a
        template that renders as though it were fine.
      security: [{ gatewayIdentity: [outreach:templates:write] }]
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '204': { description: Deleted. }
        '404': { $ref: '#/components/responses/NotFound' }

  # ── playbooks ──────────────────────────────────────────────────────────────
...
  /v1/campaigns:
    get:
      operationId: getCampaigns
      tags: [campaigns]
      summary: List campaigns
      responses:
        '200':
          description: Every campaign this tenant owns.
          content: { application/json: { schema: { type: object, properties: { data: { type: array, items: { $ref: '#/components/schemas/Campaign' } } } } } }
    post:
      operationId: createCampaign
      tags: [campaigns]
      summary: Create a campaign
      description: >
        A campaign is a playbook plus an audience plus a schedule. `context` is
        merged into every recipient's trigger payload and is subject to the
        playbook's data contract — a field the contract does not declare renders
        empty rather than leaking.
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CampaignCreate' } } }
      responses:
        '201':
          description: Created, in DRAFT.
          content: { application/json: { schema: { type: object, properties: { id: { type: string, format: uuid } } } } }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    gatewayIdentity:
      type: apiKey
      in: header
      name: x-gateway-request
      description: |
        The gateway forwards `x-gateway-request: true` plus `x-user-id`,
        `x-user-role`, `x-user-permissions` (a JSON array), `x-tenant-id` (or
        the legacy `x-medspa-id`), `x-sub-tenant-id` (or `x-location-id`) and
        `x-sender-id` (or `x-provider-id`).

        A request without `x-gateway-request` is **403**. A request with it but
        no user context is **401**. Scopes listed on an operation are the
        permissions it requires from `x-user-permissions`; the `admin` role and
        the `outreach:admin` permission bypass the check.
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string }
...
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
  responses:
    BadRequest:
      description: The request did not validate.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
    NotFound:
      description: No such resource in this tenant.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
  schemas:
    Error:
      type: object
      required: [success, error]
      properties:
        success: { type: boolean, enum: [false] }
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details: { type: object, additionalProperties: true }
        requestId: { type: string, format: uuid }
    Channel:
      type: string
      enum: [email, sms, slack, push, webhook, in_app, voice, letter]
...
    Campaign:
      type: object
      properties:
        id: { type: string, format: uuid }
        tenantId: { type: string }
        name: { type: string }
        description: { type: string, nullable: true }
        type: { type: string }
        status: { type: string, enum: [DRAFT, QUEUED, PROCESSING, PAUSED, COMPLETED, CANCELLED] }
        playbookId: { type: string, format: uuid, nullable: true }
        audienceId: { type: string, format: uuid, nullable: true }
        senderId: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CampaignCreate:
      type: object
      required: [name, playbookKey, audienceId]
      properties:
        name: { type: string }
        description: { type: string }
        type: { type: string }
        playbookKey: { type: string, description: Must be installed for this tenant. }
        audienceId: { type: string, format: uuid }
        senderId: { type: string }
        context:
          type: object
          additionalProperties: true
          description: >
            Merged into every recipient's trigger payload, then validated
            against the playbook's data contract — an undeclared field renders
            empty rather than leaking.
        schedule:
          type: object
          properties: { sendAt: { type: string, format: date-time } }
        metadata: { type: object, additionalProperties: true }
...
    RenderContext:
      type: object
      description: |
        Namespaced deliberately. `recipient`, `sender` and `tenant` are resolved
        by the engine; `context` is whatever the caller supplied, validated
        against the playbook's data contract.
      properties:
        recipient: { type: object, additionalProperties: true }
        sender: { type: object, additionalProperties: true }
        tenant: { type: object, additionalProperties: true }
        context: { type: object, additionalProperties: true }
        message: { type: object, additionalProperties: true }
    GeneratedDraft:
      type: object
      required: [draftId, content, channel, aiConfidence]
      properties:
        draftId: { type: string, format: uuid }
        content: { type: string }
        subject: { type: string }
        channel: { $ref: '#/components/schemas/Channel' }
        aiConfidence:
          type: number
          description: |
            Deterministic composite: context completeness, less 0.15 per lint
            warning. **Not** the model's self-assessment, and a heuristic rather
            than a probability.
        promptPackKey: { type: string }
        model: { type: string }
        tokensIn: { type: integer }
        tokensOut: { type: integer }
        costUsd: { type: number }
        lintWarnings: { type: array, items: { type: string } }
...
    Asset:
      type: object
      properties:
        id: { type: string, format: uuid }
        tenantId: { type: string }
        subTenantId: { type: [string, 'null'], format: uuid }
        kind: { type: string, enum: [image, document, generated] }
        url:
          type: string
          description: |
            Where the object can be fetched. Built from `ASSET_PUBLIC_BASE_URL`
            for local storage, or from `S3_PUBLIC_BASE_URL` / the bucket's
            virtual host for S3. Stored on the row, so changing the setting does
            not rewrite URLs already handed out.
        mimeType: { type: [string, 'null'] }
        size: { type: [integer, 'null'], description: Bytes. }
        metadata:
          type: object
          additionalProperties: true
          description: |
            Carries `storageKey`, `storageProvider`, `filename` (as the client
            sent it) and `uploadedBy`. A generated asset also carries `prompt`
            and `model`.
        createdAt: { type: string, format: date-time }
...